docs: Update PR template to something people might actually use
I notice that people largely ignore PR templates, *especially* big ones.
I think that shrinking this down to just the essentials will increase the
likelihood that people use it.

In particular, I want to remove things like "All reviewers approved" and
"CI build is green" that are generally already enforced by repo policies,
"Delete working branch" (there's a setting for that if we want it!), and
release instructions (should be in a central document linked from the
README).