feat!: Switch to a single Producer, wrapped in an API singleton
Purpose:

- Revisit https://github.com/openedx/event-bus-kafka/issues/16 since I
  finally figured out a clean way to have a single producer.
- Reduce the burden on future code that will need to adjust how polling
  is done (https://github.com/openedx/event-bus-kafka/issues/31) and
  maybe handle shutdown (https://github.com/openedx/event-bus-kafka/issues/11)
- Prepare for configurable implementation loading, which will need a
  singleton and getter: https://github.com/openedx/openedx-events/issues/87
- Get rid of the `sync` argument (which didn't fit the abstraction) and
  move it to a dedicated method.

Relying code should now call `get_producer().send(...)` rather than
`send_to_event_bus(...)`. The return value is an object that wraps a
`Producer` instance (not a `SerializingProducer`) and that handles the
serialization itself.

Serialization logic is moved to a cached `get_serializers(...)` that
expands upon the previous `get_serializer` function; it now returns a pair
of key and value serializers. This also acts as a patch point for
mocking. I'd like to test the serializers themselves, but they want to
talk to a server.

`send_to_event_bus` gets a shorter name (now it's just a `send` method)
and loses the `sync` keyword argument; there is instead now a
`pre_shutdown` method.