feat: Change `EVENT_BUS_KAFKA_CONSUMERS_ENABLED` default to True
This was originally intended as a rollout switch, but running the consumer
already requires invoking a management command. Defaulting to True makes
this an emergency kill-switch, which is more plausibly useful.

Also add the check to `consume_indefinitely` itself just in case it is
ever called directly rather than via the Command.

Remove ticket link that is now outdated.