feat: Log and record receiver errors the same way as other errors
This will aid in error recovery -- if we need to replay a chunk of the
topic, we'll need to have partition and offset information in the logs.
The thrown exception is given its own class in order to help with building
alerts that can discriminate between different sources of errors (business
logic unique to the consuming IDA vs. more generic kafka/avro issues).

This is part of https://github.com/openedx/event-bus-kafka/issues/62