feat: Add the ability to override offsets using timestamp.
In the case where we would want to replay events,
we would need to change the offsets in the partitions so
that we can consume from an earlier point. This allows users to pass
in a timestamp parameter that will reset the offsets to that timestamp before
starting to consume again.

https://github.com/edx/edx-arch-experiments/issues/115